Fix categories where a previously deleted category was not removed
